refactor: extract MultipleAyah from quran khatm page
